backendf/requirements.txt
Flask
flask-cors
mongoengine
python-dotenv
bcrypt
PyJWT
